refactor: move MobileFilters
